Oh no what's happening Mreow's doing oh no

+++ network.go
@@ -198,7 +198,7 @@ func (sess *Session) listenSSH() error {
 			return errors.New("Packet was too big")
 		}
 		packetLen := make([]byte, 4)
-		binary.BigEndian.PutUint16(packetLen, uint16(len(packetData)))
+		binary.BigEndian.PutUint32(packetLen, uint32(len(packetData)))
 		packetData = append(packetLen, packetData...)
 
 		fmt.Println("Done marshal")
@@ -245,7 +245,7 @@ func readProtoPacket(conn *Connection) (packets.Packet, error) {
 	if lenErr != nil {
 		return packet, lenErr
 	}
-	l := binary.BigEndian.Uint16(packetLen)
+	l := binary.BigEndian.Uint32(packetLen)
 	fmt.Println("Reading packet of length", l)
 	packetData := make([]byte, l)
 	_, dataErr := io.ReadFull(conn.conn, packetData)
